Fixes/first checkpoint

+++ tests/integral/test_delete_user.py
@@ -32,7 +32,7 @@ def test_01_delete_user_returns_204_and_deletes_user():
         session.commit()
     client = TestClient(app)
     headers = {"Authorization": "Bearer admin_token"}
-    response = client.delete(f"/users/{user_id}", headers=headers)
+    response = client.delete(f"/users/admin/{user_id}", headers=headers)
     assert response.status_code == 204
     with Session(sync_engine) as session:
         deleted_user = session.get(UserAccount, user_id)
@@ -61,7 +61,7 @@ def test_02_delete_user_with_many_tokens_deletes_all_tokens():
         session.commit()
     client = TestClient(app)
     headers = {"Authorization": "Bearer admin_token"}
-    response = client.delete(f"/users/{user_id}", headers=headers)
+    response = client.delete(f"/users/admin/{user_id}", headers=headers)
     assert response.status_code == 204
     with Session(sync_engine) as session:
         deleted_user = session.get(UserAccount, user_id)
@@ -83,28 +83,28 @@ def test_03_delete_nonexistent_user_returns_404():
     user_id = uuid.uuid4()
     client = TestClient(app)
     headers = {"Authorization": "Bearer admin_token"}
-    response = client.delete(f"/users/{user_id}", headers=headers)
+    response = client.delete(f"/users/admin/{user_id}", headers=headers)
     assert response.status_code == 404
     assert response.json() == {
         "type": "about:blank",
         "title": "Resource Not Found",
         "status": 404,
         "detail": f"Usuario con id: {user_id} no encontrado",
-        "instance": f"/users/{user_id}"
+        "instance": f"/users/admin/{user_id}"
     }
     app.dependency_overrides = {}
 
 def test_04_delete_user_without_admin_token_returns_401():
     user_id = uuid.uuid4()
     client = TestClient(app)
-    response = client.delete(f"/users/{user_id}")
+    response = client.delete(f"/users/admin/{user_id}")
     assert response.status_code == 401
     assert response.json() == {
         "type": "about:blank",
         "title": "Authentication Error",
         "status": 401,
         "detail": "Token de autenticación invalido o no proporcionado",
-        "instance": f"/users/{user_id}"
+        "instance": f"/users/admin/{user_id}"
     }
 
 def test_05_delete_user_with_non_admin_token_returns_401():
@@ -112,13 +112,36 @@ def test_05_delete_user_with_non_admin_token_returns_401():
     user_id = uuid.uuid4()
     client = TestClient(app)
     headers = {"Authorization": f"Bearer {user_id}"}
-    response = client.delete(f"/users/{user_id}", headers=headers)
+    response = client.delete(f"/users/admin/{user_id}", headers=headers)
     assert response.status_code == 401
     assert response.json() == {
         "type": "about:blank",
         "title": "Authentication Error",
         "status": 401,
         "detail": "Se requiere privilegios de administrador",
-        "instance": f"/users/{user_id}"
+        "instance": f"/users/admin/{user_id}"
+    }
+    app.dependency_overrides = {}
+
+def test_06_when_delete_user_then_the_user_request_are_invalid():
+    app.dependency_overrides[require_admin] = override_require_admin
+    user_id = uuid.uuid4()
+    with Session(sync_engine) as session:
+        user = UserAccount(id=user_id, email="test@example.com", password="password")
+        session.add(user)
+        session.commit()
+    client = TestClient(app)
+    headers = {"Authorization": "Bearer admin_token"}
+    response = client.delete(f"/users/admin/{user_id}", headers=headers)
+    assert response.status_code == 204
+    app.dependency_overrides[get_jwt_payload] = lambda: {"user_id": str(user_id), "role": "listener"}
+    response = client.get("/users/me", headers={"Authorization": f"Bearer {user_id}"})
+    assert response.status_code == 401
+    assert response.json() == {
+        "type": "about:blank",
+        "title": "Authentication Error",
+        "status": 401,
+        "detail": "Usuario no encontrado",
+        "instance": "/users/me"
     }
     app.dependency_overrides = {}
